chore: remove forwarding
